--- docs/source/estimator_overview/estimator_overview.html
@@ -1,127 +1,0 @@
-<!DOCTYPE html>
-<html>
-
-<head>
-  <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.4.1/jquery.min.js">
-  </script>
-
-  <style>
-    html {
-      font-family: "Lucida Sans Unicode", "Lucida Grande", Sans-Serif;
-    }
-
-    input {
-      background-image: url('https://www.w3schools.com/css/searchicon.png');
-      /* Add a search icon to input */
-      /*background-position: 10px 12px;*/
-      /* Position the search icon */
-      background-repeat: no-repeat;
-      /* Do not repeat the icon image */
-      width: 30%;
-      /* Full-width */
-      font-size: 16px;
-      /* Increase font-size */
-      padding: 2px 20px 4px 30px;
-      /* Add some padding */
-      border: 1px solid #ddd;
-      /* Add a grey border */
-      margin-bottom: 10px;
-      /* Add some space below the input */
-      margin-left: 30%;
-    }
-
-    table {
-      margin: 0 auto;
-      width: 100%;
-      clear: both;
-      border-spacing: 0;
-      border: 1px solid white;
-      overflow: auto;
-    }
-
-    th,
-    td {
-      border: 1px solid #ddd;
-      color: #6678b1;
-      text-align: left;
-      height: 50px;
-      margin: 0;
-      font-size: 12px;
-      /* Increase font-size */
-      outline: none;
-    }
-
-    thead th {
-      border-top: 1px solid #333131;
-      border-bottom: 1px solid #333131;
-      text-align: center;
-      font-size: 15px;
-      /* Increase font-size */
-      padding: 5px;
-      font-weight: bold;
-      outline: none;
-      color: #039 !important;
-    }
-
-    .dataframe tr:nth-child(even) {
-      background-color: #f2f2f2;
-    }
-  </style>
-</head>
-
-<body>
-<label for="myInput"></label><input type="text" id="myInput" placeholder="Search for names..">
-  <br>
-  <br>
-
-  <div
-          w3-include-html="../source/estimator_overview/estimator_overview_table.html"></div>
-</body>
-
-<script>
-  $(document).ready(function () {
-    $("#myInput").on("keyup", function () {
-      var value = $(this).val().toLowerCase();
-      $("tr").filter(function () {
-        $(this).toggle($(this).text()
-          .toLowerCase().indexOf(value) > -1)
-      });
-    });
-  });
-</script>
-
-<script>
-  function includeHTML() {
-    var z, i, elmnt, file, xhttp;
-    /*loop through a collection of all HTML elements:*/
-    z = document.getElementsByTagName("*");
-    for (i = 0; i < z.length; i++) {
-      elmnt = z[i];
-      /*search for elements with a certain atrribute:*/
-      file = elmnt.getAttribute("w3-include-html");
-      if (file) {
-        /*make an HTTP request using the attribute value as the file name:*/
-        xhttp = new XMLHttpRequest();
-        xhttp.onreadystatechange = function () {
-          if (this.readyState === 4) {
-            if (this.status === 200) { elmnt.innerHTML = this.responseText; }
-            if (this.status === 404) { elmnt.innerHTML = "Page not found."; }
-            /*remove the attribute, and call this function once more:*/
-            elmnt.removeAttribute("w3-include-html");
-            includeHTML();
-          }
-        }
-        xhttp.open("GET", file, true);
-        xhttp.send();
-        /*exit the function:*/
-        return;
-      }
-    }
-  };
-</script>
-
-<script>
-  includeHTML();
-</script>
-
-</html>
